Extract code for reading a line at a time, from head/tail/wc builtins
fileLines() is slightly different from what these utilities were doing:
it exposes the lines as a generator so that they can be iterated.
Internally it still reads the entire input files, but in future it can
be modified to only read one line at a time, without having to modify
the code that uses it.
